use u16 memsize

+++ src/test/dataTest.h
@@ -9,7 +9,7 @@ struct Class_dataTest
     void *context;
     int32_t a;
     int32_t b;
-    uint32_t memSize;
+    uint16_t memSize;
 
     /* operation */
     void (*deinit)(dataTest_t *self);
